Changed: split engine.model.Selection to Selection and DocumentSelection.
