+++ src/model/selection.js
@@ -7,21 +7,13 @@
 
 import Position from './position.js';
 import Range from './range.js';
-import LiveRange from './liverange.js';
 import EmitterMixin from '../../utils/emittermixin.js';
-import CharacterProxy from './characterproxy.js';
 import CKEditorError from '../../utils/ckeditorerror.js';
-import toMap from '../../utils/tomap.js';
 import mix from '../../utils/mix.js';
 
-const storePrefix = 'selection:';
-
 /**
- * Represents a selection that is made on nodes in {@link engine.model.Document}. `Selection` instance is
- * created by {@link engine.model.Document}. You should not need to create an instance of `Selection`.
- *
- * Keep in mind that selection always contains at least one range. If no ranges has been added to selection or all ranges
- * got removed from selection, the selection will be reset to contain {@link engine.model.Selection#_getDefaultRange the default range}.
+ * `Selection` is a group of {@link engine.model.Range ranges} which has a direction specified by
+ * {@link engine.model.Selection#anchor anchor} and {@link engine.model.Selection#focus focus}.
  *
  * @memberOf engine.model
  */
@@ -32,14 +24,6 @@ export default class Selection {
 	 * @param {engine.model.Document} document Document which owns this selection.
 	 */
 	constructor( document ) {
-		/**
-		 * List of attributes set on current selection.
-		 *
-		 * @protected
-		 * @member {Map} engine.model.Selection#_attrs
-		 */
-		this._attrs = new Map();
-
 		/**
 		 * Document which owns this selection.
 		 *
@@ -60,7 +44,7 @@ export default class Selection {
 		 * Stores all ranges that are selected.
 		 *
 		 * @private
-		 * @member {Array.<engine.model.LiveRange>} engine.model.Selection#_ranges
+		 * @member {Array.<engine.model.Range>} engine.model.Selection#_ranges
 		 */
 		this._ranges = [];
 	}
@@ -69,10 +53,9 @@ export default class Selection {
 	 * Selection anchor. Anchor may be described as a position where the selection starts. Together with
 	 * {@link engine.model.Selection#focus} they define the direction of selection, which is important
 	 * when expanding/shrinking selection. Anchor is always the start or end of the most recent added range.
-	 * It may be a bit unintuitive when there are multiple ranges in selection.
 	 *
 	 * @see engine.model.Selection#focus
-	 * @type {engine.model.LivePosition}
+	 * @type {engine.model.Position}
 	 */
 	get anchor() {
 		let range = this._ranges.length ? this._ranges[ this._ranges.length - 1 ] : this._getDefaultRange();
@@ -84,7 +67,7 @@ export default class Selection {
 	 * Selection focus. Focus is a position where the selection ends.
 	 *
 	 * @see engine.model.Selection#anchor
-	 * @type {engine.model.LivePosition}
+	 * @type {engine.model.Position}
 	 */
 	get focus() {
 		let range = this._ranges.length ? this._ranges[ this._ranges.length - 1 ] : this._getDefaultRange();
@@ -130,12 +113,12 @@ export default class Selection {
 	}
 
 	/**
-	 * Adds a range to the selection. Added range is copied and converted to {@link engine.model.LiveRange}. This means
-	 * that passed range is not saved in the Selection instance and you can safely operate on it.
+	 * Adds a range to the selection. Added range is copied. This means that passed range is not saved in `Selection`
+	 * instance and operating on it will not change `Selection` state.
 	 *
 	 * Accepts a flag describing in which way the selection is made - passed range might be selected from
-	 * {@link engine.model.Range#start} to {@link engine.model.Range#end} or from {@link engine.model.Range#end}
-	 * to {@link engine.model.Range#start}. The flag is used to set {@link engine.model.Selection#anchor} and
+	 * {@link engine.model.Range#start start} to {@link engine.model.Range#end end} or from {@link engine.model.Range#end end}
+	 * to {@link engine.model.Range#start start}. The flag is used to set {@link engine.model.Selection#anchor} and
 	 * {@link engine.model.Selection#focus} properties.
 	 *
 	 * @fires engine.model.Selection#change:range
@@ -151,16 +134,7 @@ export default class Selection {
 	}
 
 	/**
-	 * Unbinds all events previously bound by this selection or objects created by this selection.
-	 */
-	destroy() {
-		for ( let i = 0; i < this._ranges.length; i++ ) {
-			this._ranges[ i ].detach();
-		}
-	}
-
-	/**
-	 * Returns an iterator that contains copies of all ranges added to the selection.
+	 * Returns an iterator that iterates over copies of selection ranges.
 	 *
 	 * @returns {Iterator.<engine.model.Range>}
 	 */
@@ -211,7 +185,6 @@ export default class Selection {
 	 * @fires engine.model.Selection#change:range
 	 */
 	removeAllRanges() {
-		this.destroy();
 		this._ranges = [];
 
 		this.fire( 'change:range' );
@@ -228,7 +201,6 @@ export default class Selection {
 	 * or backward - from end to start (`true`). Defaults to `false`.
 	 */
 	setRanges( newRanges, isLastBackward ) {
-		this.destroy();
 		this._ranges = [];
 
 		for ( let i = 0; i < newRanges.length; i++ ) {
@@ -277,8 +249,7 @@ export default class Selection {
 		const anchor = this.anchor;
 
 		if ( this._ranges.length ) {
-			// TODO Replace with _popRange, so child classes can override this (needed for #329).
-			this._ranges.pop().detach();
+			this._popRange();
 		}
 
 		if ( newFocus.compareWith( anchor ) == 'BEFORE' ) {
@@ -289,95 +260,14 @@ export default class Selection {
 	}
 
 	/**
-	 * Removes all attributes from the selection.
+	 * Checks if given range intersects with ranges that are already in the selection. Throws an error if it does.
+	 * This method is extracted from {@link engine.model.Selection#_pushRange } so it is easier to override it.
 	 *
-	 * @fires engine.model.Selection#change:attribute
+	 * @param {engine.model.Range} range Range to check.
+	 * @protected
 	 */
-	clearAttributes() {
-		this._attrs.clear();
-		this._setStoredAttributesTo( new Map() );
-
-		this.fire( 'change:attribute' );
-	}
-
-	/**
-	 * Gets an attribute value for given key or undefined it that attribute is not set on selection.
-	 *
-	 * @param {String} key Key of attribute to look for.
-	 * @returns {*} Attribute value or null.
-	 */
-	getAttribute( key ) {
-		return this._attrs.get( key );
-	}
-
-	/**
-	 * Returns iterator that iterates over this selection attributes.
-	 *
-	 * @returns {Iterable.<*>}
-	 */
-	getAttributes() {
-		return this._attrs[ Symbol.iterator ]();
-	}
-
-	/**
-	 * Checks if the selection has an attribute for given key.
-	 *
-	 * @param {String} key Key of attribute to check.
-	 * @returns {Boolean} `true` if attribute with given key is set on selection, `false` otherwise.
-	 */
-	hasAttribute( key ) {
-		return this._attrs.has( key );
-	}
-
-	/**
-	 * Removes an attribute with given key from the selection.
-	 *
-	 * @fires engine.model.Selection#change:attribute
-	 * @param {String} key Key of attribute to remove.
-	 */
-	removeAttribute( key ) {
-		this._attrs.delete( key );
-		this._removeStoredAttribute( key );
-
-		this.fire( 'change:attribute' );
-	}
-
-	/**
-	 * Sets attribute on the selection. If attribute with the same key already is set, it overwrites its values.
-	 *
-	 * @fires engine.model.Selection#change:attribute
-	 * @param {String} key Key of attribute to set.
-	 * @param {*} value Attribute value.
-	 */
-	setAttribute( key, value ) {
-		this._attrs.set( key, value );
-		this._storeAttribute( key, value );
-
-		this.fire( 'change:attribute' );
-	}
-
-	/**
-	 * Removes all attributes from the selection and sets given attributes.
-	 *
-	 * @fires engine.model.Selection#change:attribute
-	 * @param {Iterable|Object} attrs Iterable object containing attributes to be set.
-	 */
-	setAttributesTo( attrs ) {
-		this._attrs = toMap( attrs );
-		this._setStoredAttributesTo( this._attrs );
-
-		this.fire( 'change:attribute' );
-	}
-
-	/**
-	 * Converts given range to {@link engine.model.LiveRange} and adds it to internal ranges array. Throws an error
-	 * if given range is intersecting with any range that is already stored in this selection.
-	 *
-	 * @private
-	 * @param {engine.model.Range} range Range to add.
-	 */
-	_pushRange( range ) {
-		for ( let i = 0; i < this._ranges.length ; i++ ) {
+	_checkRange( range ) {
+		for ( let i = 0; i < this._ranges.length; i++ ) {
 			if ( range.isIntersecting( this._ranges[ i ] ) ) {
 				/**
 				 * Trying to add a range that intersects with another range from selection.
@@ -392,173 +282,27 @@ export default class Selection {
 				);
 			}
 		}
-
-		this._ranges.push( LiveRange.createFromRange( range ) );
 	}
 
 	/**
-	 * Iterates through all attributes stored in current selection's parent.
-	 *
-	 * @returns {Iterable.<*>}
-	 */
-	*_getStoredAttributes() {
-		const selectionParent = this.getFirstPosition().parent;
-
-		if ( this.isCollapsed && selectionParent.getChildCount() === 0 ) {
-			for ( let attr of selectionParent.getAttributes() ) {
-				if ( attr[ 0 ].indexOf( storePrefix ) === 0 ) {
-					const realKey = attr[ 0 ].substr( storePrefix.length );
-
-					yield [ realKey, attr[ 1 ] ];
-				}
-			}
-		}
-	}
-
-	/**
-	 * Removes attribute with given key from attributes stored in current selection's parent node.
-	 *
-	 * @private
-	 * @param {String} key Key of attribute to remove.
-	 */
-	_removeStoredAttribute( key ) {
-		const selectionParent = this.getFirstPosition().parent;
-
-		if ( this.isCollapsed && selectionParent.getChildCount() === 0 ) {
-			const storeKey = Selection._getStoreAttributeKey( key );
-
-			this._document.enqueueChanges( () => {
-				this._document.batch().removeAttr( storeKey, selectionParent );
-			} );
-		}
-	}
-
-	/**
-	 * Stores given attribute key and value in current selection's parent node if the selection is collapsed and
-	 * the parent node is empty.
-	 *
-	 * @private
-	 * @param {String} key Key of attribute to set.
-	 * @param {*} value Attribute value.
-	 */
-	_storeAttribute( key, value ) {
-		const selectionParent = this.getFirstPosition().parent;
-
-		if ( this.isCollapsed && selectionParent.getChildCount() === 0 ) {
-			const storeKey = Selection._getStoreAttributeKey( key );
-
-			this._document.enqueueChanges( () => {
-				this._document.batch().setAttr( storeKey, value, selectionParent );
-			} );
-		}
-	}
-
-	/**
-	 * Sets selection attributes stored in current selection's parent node to given set of attributes.
-	 *
-	 * @param {Iterable|Object} attrs Iterable object containing attributes to be set.
-	 * @private
-	 */
-	_setStoredAttributesTo( attrs ) {
-		const selectionParent = this.getFirstPosition().parent;
-
-		if ( this.isCollapsed && selectionParent.getChildCount() === 0 ) {
-			this._document.enqueueChanges( () => {
-				const batch = this._document.batch();
-
-				for ( let attr of this._getStoredAttributes() ) {
-					const storeKey = Selection._getStoreAttributeKey( attr[ 0 ] );
-
-					batch.removeAttr( storeKey, selectionParent );
-				}
-
-				for ( let attr of attrs ) {
-					const storeKey = Selection._getStoreAttributeKey( attr[ 0 ] );
-
-					batch.setAttr( storeKey, attr[ 1 ], selectionParent );
-				}
-			} );
-		}
-	}
-
-	/**
-	 * Updates this selection attributes based on it's position in the model.
+	 * Removes most recently added range from the selection.
 	 *
 	 * @protected
 	 */
-	_updateAttributes() {
-		const position = this.getFirstPosition();
-		const positionParent = position.parent;
+	_popRange() {
+		this._ranges.pop();
+	}
 
-		let attrs = null;
-
-		if ( !this.isCollapsed ) {
-			// 1. If selection is a range...
-			const range = this.getFirstRange();
-
-			// ...look for a first character node in that range and take attributes from it.
-			for ( let item of range ) {
-				// This is not an optimal solution because of https://github.com/ckeditor/ckeditor5-engine/issues/454.
-				// It can be done better by using `break;` instead of checking `attrs === null`.
-				if ( item.type == 'TEXT' && attrs === null ) {
-					attrs = item.item.getAttributes();
-				}
-			}
-		} else {
-			// 2. If the selection is a caret or the range does not contain a character node...
-
-			const nodeBefore = positionParent.getChild( position.offset - 1 );
-			const nodeAfter = positionParent.getChild( position.offset );
-
-			// ...look at the node before caret and take attributes from it if it is a character node.
-			attrs = getAttrsIfCharacter( nodeBefore );
-
-			// 3. If not, look at the node after caret...
-			if ( !attrs ) {
-				attrs = getAttrsIfCharacter( nodeAfter );
-			}
-
-			// 4. If not, try to find the first character on the left, that is in the same node.
-			if ( !attrs ) {
-				let node = nodeBefore;
-
-				while ( node && !attrs ) {
-					node = node.previousSibling;
-					attrs = getAttrsIfCharacter( node );
-				}
-			}
-
-			// 5. If not found, try to find the first character on the right, that is in the same node.
-			if ( !attrs ) {
-				let node = nodeAfter;
-
-				while ( node && !attrs ) {
-					node = node.nextSibling;
-					attrs = getAttrsIfCharacter( node );
-				}
-			}
-
-			// 6. If not found, selection should retrieve attributes from parent.
-			if ( !attrs ) {
-				attrs = this._getStoredAttributes();
-			}
-		}
-
-		if ( attrs ) {
-			this._attrs = new Map( attrs );
-		} else {
-			this.clearAttributes();
-		}
-
-		function getAttrsIfCharacter( node ) {
-			if ( node instanceof CharacterProxy ) {
-				return node.getAttributes();
-			}
-
-			return null;
-		}
-
-		this.fire( 'change:attribute' );
+	/**
+	 * Adds given range to internal {@link engine.model.Selection#_ranges ranges array}. Throws an error
+	 * if given range is intersecting with any range that is already stored in this selection.
+	 *
+	 * @protected
+	 * @param {engine.model.Range} range Range to add.
+	 */
+	_pushRange( range ) {
+		this._checkRange( range );
+		this._ranges.push( Range.createFromRange( range ) );
 	}
 
 	/**
@@ -584,29 +328,12 @@ export default class Selection {
 
 		return new Range( position, position );
 	}
-
-	/**
-	 * Generates and returns an attribute key for selection attributes store, basing on original attribute key.
-	 *
-	 * @param {String} key Attribute key to convert.
-	 * @returns {String} Converted attribute key, applicable for selection store.
-	 */
-	static _getStoreAttributeKey( key ) {
-		return storePrefix + key;
-	}
 }
 
 mix( Selection, EmitterMixin );
 
 /**
- * Fired whenever selection ranges are changed through {@link engine.model.Selection Selection API}. Not fired when
- * {@link engine.model.LiveRange live ranges} inserted in selection change because of Tree Model changes.
+ * Fired whenever selection ranges are changed through {@link engine.model.Selection Selection API}.
  *
  * @event engine.model.Selection#change:range
- */
-
-/**
- * Fired whenever selection attributes are changed.
- *
- * @event engine.model.Selection#change:attribute
  */
